requirements.txt
bitarray
datasets
gitpython
spacy[cuda12x]
spacy-transformers
tqdm
transformers
torch
pytorch-lightning
hkkang_utils
tensorboard
beir
unidecode
matplotlib
hydra-core
lz4
peft
wget